allow simultaneous launching of 2+ worker processes per host
One of the logic changes in PR #65 broke the ability to simultaneously launch
more than 1 worker process for a given topology.  The cause of the breakage
was intentionally avoiding inclusion of the executor's resources into the
ExecutorInfo structure associated with the mesos tasks (storm workers).
Notably, the avoidance is only triggered for tasks other than the 1st one
that potentially launches the executor.

This is problematic because the mesos-master rejects tasks whose ExecutorInfo
isn't identical to other tasks under the same executor.  Notably, since the
executor is already running for subsequent tasks, the resources that are
added to these subsequent tasks' ExecutorInfo aren't actually used, so there
is no advantage in attempting to avoid their inclusion.

FFR, this is the commit with that change:
* af8c49beac04b530c33c1401c829caaa8e368a35

After this fix I was able to instantly launch 3 workers for a topology
on the same mesos-slave host.