fix issue with vagrant start-nimbus.sh script (#124)
After we increased the package version name with all the useful info about
mesos and storm version, the process names are now too long for us to be
able to see the whole command in `ps` output (Linux limits the process length
to 4096 bytes by default).  So we cannot search for the correct daemon name
since it is cut off.  The brutal workaround is to kill all processes with
'storm' in their name.