Set the MesosSupervisor's storm.local.dir config to prevent conflicts between topologies
Without this change, we observed one supervisor killing all other topology's workers on a host,
which was being triggered because we were explicitly setting the MesosNimbus's storm.local.dir
to a hardcoded path.  If you don't set storm.local.dir in MesosNimbus then you would not be
affected by this problem.

See the comments in the code for the details on how this happens.