update the DOCKER_REPO variable to be the full repo path
This change avoids the assumption that the target docker repo will be `foo/storm`,
where only the `foo` portion is controlled by `DOCKER_REPO`.
With this fix it could be `foo/storm-mesos` for example.

The current `DOCKER_REPO` variable isn't referring to the full repo "slug", instead
it just references the user/organization portion of the slug, and assumes the project
name will always be `storm`.  That is not necessarily the case, for example in my fork
of `mesos/storm` I call the project `erikdw/storm-mesos` instead of just `erikdw/storm`.
That is done to avoid a conflict with the actual storm project (`apache/storm`), which
I also have forked under my name (`erikdw/storm`).