adjusts tests to account for some corner cases that exist for the time being
1. Our floating point tabulations in AggregatedOffers can lead to confusing results
where a 1.0 becomes 0.999999999996 for example.  We add a fudge factor by adding 0.01
more CPU resources in one test to work around this until the following issue is fixed:
  * https://github.com/mesos/storm/issues/161
2. Since we are disabling the ability of allSlotsAvailableForScheduling to account
for supervisors already existing when calculating resource needs, we disable the test
that is validating that behavior.  That will be reenabled once we fix this issue:
* https://github.com/mesos/storm/issues/160

Also a few cosmetic-ish changes:
1. Use more standard hostnames with a domain of "example.org", a standard domain
for documentation.  See http://example.org
2. Rearrange/renumber some of the offers to prevent confusion.