Change LocalFileServer to allow respecting storm.yaml's nimbus.host
Prior to this the only way to configure the URL that LocalFileServer listens
to is via MESOS_NIMBUS_HOST environment variable.  That differs from the
behavior of the core MesosNimbus which supports that env variable, as well
as the nimbus.host field in the storm.yaml config.  So update the code to
make them equally configurable.